Move function templates for defining numerical operations to header
